project/requirements.txt
fastapi==0.60.1
pandas==1.1.3
plotly==4.12.0
uvicorn==0.11.8
sodapy==2.1.0
scikit-learn==0.23.2
requests==2.24.0
python-dotenv==0.15.0
psycopg2-binary==2.8.6
pytz==2020.5
